namespace and a small abstraction between result and it's components
