+++ src/parsed_component.hpp
@@ -1,9 +1,10 @@
 #pragma once
 
-#include <iostream>
 #include <ctime>
-#include <unordered_map>
+#include <iostream>
 #include <string>
+#include <unordered_map>
+
 #include "src/utils/utils.hpp"
 
 typedef std::unordered_map<std::string, std::pair<bool, int> > Components;
@@ -52,40 +53,5 @@ public:
     void setTimeZoneOffset(int);
 
 };
-
-
-class parse::ParsedResult {
-private:
-    bool __end;
-protected:
-    posix_time::ptime anchor;
-    unsigned index;
-    std::string text; // todo: add accessor for text to aid testing
-    utils::Tags tags;
-
-public:
-    ParsedResult();
-    ParsedResult(posix_time::ptime, long, std::string);
-    ParsedResult(const ParsedResult& pr);
-    ~ParsedResult();
-    // ParsedResult& operator=(ParsedResult&);
-
-    ParsedComponents startDate, endDate;    // todo: make protected & make accessors and mutators
-
-    bool isEmpty() const;
-    void setTag(utils::Modifiers);
-    bool hasPossibleDates();
-    std::string toDate();
-    unsigned getIndex() const;
-    bool getTag(utils::Modifiers);
-    void setIndex(int idx);
-    void setText(std::string);
-    size_t textLength() const;
-    bool end() const;
-    void makeEndDateValid();
-    parse::ParsedResult& operator=(parse::ParsedResult);
-};
-
-typedef std::vector<parse::ParsedResult> Result;
