+++ src/parsers/parsers.hpp
@@ -5,43 +5,44 @@
 #include <iostream>
 #include <string_view>
 
-#include "src/result.hpp"
+#include "src/parsed_result.hpp"
+
 #include "src/utils/utils.hpp"
 
 namespace parser
 {
-    static const unsigned short FULL_MATCH = 0;
+static const unsigned short FULL_MATCH = 0;
 
-    class Parser
+class Parser
+{
+protected:
+    bool strictMode;
+    std::regex pattern;
+
+    std::shared_ptr<Parser> _next;
+
+public:
+    Parser();
+
+    //Parser(bool, std::regex);       // strictMode, pattern
+    virtual
+    std::regex getPattern() const { return std::regex(); };
+
+    virtual
+    parse::ParsedResult extract(std::string&,
+                                const std::smatch&,
+                                const posix_time::ptime&,
+                                long)
     {
-    protected:
-        bool strictMode;
-        std::regex pattern;
+        parse::ParsedResult tmp{};
+        return tmp;
+    }
 
-        std::shared_ptr<Parser> _next;
+    void execute(std::string &, posix_time::ptime &, parse::Result&);
 
-    public:
-        Parser();
+    const
+    std::shared_ptr<Parser>& chain(const std::shared_ptr<Parser>&);
 
-        //Parser(bool, std::regex);       // strictMode, pattern
-        virtual
-        std::regex getPattern() const { return std::regex(); };
-
-        virtual
-        parse::ParsedResult extract(std::string&,
-                                    const std::smatch&,
-                                    const posix_time::ptime&,
-                                    long)
-        {
-            parse::ParsedResult tmp{};
-            return tmp;
-        }
-
-        void execute(std::string &, posix_time::ptime &, Result&);
-
-        const
-        std::shared_ptr<Parser>& chain(const std::shared_ptr<Parser>&);
-
-        ~Parser();
-    };
+    ~Parser();
+};
 }
