+++ src/refiners/en/ENMergeDateAndTimeRefiner.hpp
@@ -33,18 +33,18 @@ public:
          */
         parse::ParsedComponents dateTimeComp{dateComp};
 
-        if(timeComp.isCertain("hour")) {
+        if (timeComp.isCertain("hour"))
+        {
             dateTimeComp.setHour(timeComp.getHour());
             dateTimeComp.setMinute(timeComp.getMinute());
 
-            if(timeComp.isCertain("sec")) {
+            if (timeComp.isCertain("sec"))
                 dateTimeComp.setSeconds(timeComp.getSeconds());
-            }
-            else {
+            else
                 dateTimeComp.implyComponent("sec", timeComp.getSeconds());
-            }
-        } 
-        else {
+        }
+        else
+        {
             dateTimeComp.implyComponent("hour", timeComp.getHour());
             dateTimeComp.implyComponent("min", timeComp.getMinute());
             dateTimeComp.implyComponent("sec", timeComp.getSeconds());
@@ -64,7 +64,7 @@ public:
         parse::ParsedComponents startTime{timeResult.startDate};
         parse::ParsedComponents startDateTime{mergeDateTimeComponent(startDate, startTime)};
 
-        if(dateResult.end() or timeResult.end()) {
+        if (dateResult.end() or timeResult.end()) {
             parse::ParsedComponents endDate, endTime, endDateTime;
 
             dateResult.end() ? endDate = dateResult.endDate : endDate = dateResult.startDate;
@@ -72,9 +72,9 @@ public:
 
             endDateTime = mergeDateTimeComponent(endDate, endTime);
 
-            if(!dateResult.end() and endDateTime.date() < startDateTime.date()) {
+            if (!dateResult.end() and endDateTime.date() < startDateTime.date()) {
                 // Ex. 9pm - 1am
-                if(endDateTime.isCertain("mday"))
+                if (endDateTime.isCertain("mday"))
                     endDateTime.set_mDay(endDateTime.get_mDay() + 1);
                 else
                     endDateTime.implyComponent("mday", endDateTime.get_mDay() + 1);
@@ -92,31 +92,31 @@ public:
         dateResult.setText(text.substr(startIndex, endIndex - startIndex));
 
         for(auto &mod: utils::ModList) {
-            if(timeResult.getTag(mod))
+            if (timeResult.getTag(mod))
                 dateResult.setTag(mod);
         }
         dateResult.setTag(utils::ENMergeDateAndTimeRefiner);
         return dateResult;
     }
 
-    Result refine(Result r, std::string t) override {
+    parse::Result refine(parse::Result r, std::string t) override {
         /**
          * @brief Merge a date-only result and a time-only result into one combined result
          * @param r: list of results from parsers
          * @param t: complete original text
          * @return list of parse::ParsedResults
          */
-        if(r.size() < 2) // no second component to merge
+        if (r.size() < 2) // no second component to merge
             return r;
 
         parse::ParsedResult prev, curr, tmp;
-        Result merged_result;
+        parse::Result merged_result;
 
         for(unsigned i=1; i<r.size(); i++) {
             curr = r[i];
             prev = r[i-1];
 
-            if(isDateOnly(prev) and isTimeOnly(curr) and isAbleToMerge(t, prev, curr)) {
+            if (isDateOnly(prev) and isTimeOnly(curr) and isAbleToMerge(t, prev, curr)) {
                 prev = mergeResult(t, prev, curr);
                 try {
                     curr = r.at(i+1);
@@ -126,7 +126,7 @@ public:
                 }
                 i++;
             }
-            else if(isDateOnly(curr) and isTimeOnly(prev) and isAbleToMerge(t, prev, curr)) {
+            else if (isDateOnly(curr) and isTimeOnly(prev) and isAbleToMerge(t, prev, curr)) {
                 prev = mergeResult(t, curr, prev);
                 try {
                     curr = r.at(i+1);
@@ -138,7 +138,7 @@ public:
             }
             merged_result.push_back(prev);
         }
-        if(!curr.isEmpty()) {
+        if (!curr.isEmpty()) {
             merged_result.push_back(curr);
         }
